+++ ocean_data_parser/__init__.py
@@ -3,5 +3,7 @@
 import re
 from pathlib import Path
 
-PARSERS = re.findall('parser = "(.*)"', (Path(__file__).parent / "read.py").read_text())
+PARSERS = re.findall(
+    'parser = "(.*)"', (Path(__file__).parent / "read.py").read_text()
+) + ["dfo.odf.odf", "dfo.odf.as_qo_odf"]
 __version__ = "0.7.0"