+++ ocean_data_parser/parsers/dfo/odf_source/attributes.py
@@ -305,6 +305,12 @@ def _generate_instrument_attributes(odf_header, instrument_manufacturer_header=N
 
 
 def _generate_title_from_global_attributes(attributes):
+    org_inst = [
+        item
+        for item in [attributes.get("organization"), attributes.get("institution")]
+        if item
+    ]
+
     title = (
         f"{attributes['odf_data_type']} profile data collected "
         + (
@@ -312,8 +318,8 @@ def _generate_title_from_global_attributes(attributes):
             if "platform" in attributes
             else ""
         )
-        + f"by {attributes['organization']} {attributes['institution']} "
-        + f"on the {attributes['cruise_name'].title()} "
+        + (f" by {' '.join(org_inst) if org_inst else ''}")
+        + f" on the {attributes['cruise_name'].title()} "
     )
     if (
         pd.notna(attributes["cruise_start_date"])